--- src/main/java/agaig/justeat/domain/MedicalHelth.java
@@ -1,5 +1,0 @@
-package agaig.justeat.domain;
-
-public class MedicalHelth {
-
-}
