+++ src/main/java/agaig/justeat/domain/MemberHealth.java
@@ -4,8 +4,6 @@ public class MemberHealth {
     private Long member_id;
     private String height;
     private String weight;
-    private String sex;
-    private String birth;
     private String goal;
     private String protein;
     private String carbo;
@@ -33,22 +31,6 @@ public class MemberHealth {
 
     public void setWeight(String weight) {
         this.weight = weight;
-    }
-
-    public String getSex() {
-        return sex;
-    }
-
-    public void setSex(String sex) {
-        this.sex = sex;
-    }
-
-    public String getBirth() {
-        return birth;
-    }
-
-    public void setBirth(String birth) {
-        this.birth = birth;
     }
 
     public String getGoal() {
